requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
pymongo==4.7.2
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.1
email-validator==2.1.0.post1
